Introduce AbstractMcsException
